# frontend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Document Archive API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/archive/sets": {
      "get": {
        "summary": "Get Archive Sets",
        "operationId": "get_archive_sets_api_archive_sets_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "제목/설명/태그 부분일치 검색",
              "title": "Q"
            },
            "description": "제목/설명/태그 부분일치 검색"
          },
          {
            "name": "include_unmapped",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "set:* 태그가 없는 문서 포함 여부",
              "default": true,
              "title": "Include Unmapped"
            },
            "description": "set:* 태그가 없는 문서 포함 여부"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          },
          {
            "name": "document_limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 5,
              "description": "세트별 문서 그룹 최대 반환 수",
              "default": 50,
              "title": "Document Limit"
            },
            "description": "세트별 문서 그룹 최대 반환 수"
          },
          {
            "name": "revision_limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "문서 그룹별 리비전 최대 반환 수",
              "default": 20,
              "title": "Revision Limit"
            },
            "description": "문서 그룹별 리비전 최대 반환 수"
          },
          {
            "name": "max_documents_scanned",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20000,
              "minimum": 100,
              "description": "요청당 최대 스캔 문서 수",
              "default": 5000,
              "title": "Max Documents Scanned"
            },
            "description": "요청당 최대 스캔 문서 수"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchiveSetsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArchiveSetDocumentNode": {
        "properties": {
          "document_key": {
            "type": "string",
            "title": "Document Key"
          },
          "display_title": {
            "type": "string",
            "title": "Display Title"
          },
          "latest_event_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Event Date"
          },
          "revision_count": {
            "type": "integer",
            "title": "Revision Count"
          },
          "needs_review_count": {
            "type": "integer",
            "title": "Needs Review Count"
          },
          "kinds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Kinds"
          },
          "revisions": {
            "items": {
              "$ref": "#/components/schemas/ArchiveSetRevisionItem"
            },
            "type": "array",
            "title": "Revisions"
          },
          "has_more_revisions": {
            "type": "boolean",
            "title": "Has More Revisions",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "document_key",
          "display_title",
          "revision_count",
          "needs_review_count"
        ],
        "title": "ArchiveSetDocumentNode"
      },
      "ArchiveSetNode": {
        "properties": {
          "set_key": {
            "type": "string",
            "title": "Set Key"
          },
          "set_label": {
            "type": "string",
            "title": "Set Label"
          },
          "latest_event_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Event Date"
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count"
          },
          "revision_count": {
            "type": "integer",
            "title": "Revision Count"
          },
          "needs_review_count": {
            "type": "integer",
            "title": "Needs Review Count"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/ArchiveSetDocumentNode"
            },
            "type": "array",
            "title": "Documents"
          },
          "has_more_documents": {
            "type": "boolean",
            "title": "Has More Documents",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "set_key",
          "set_label",
          "document_count",
          "revision_count",
          "needs_review_count"
        ],
        "title": "ArchiveSetNode"
      },
      "ArchiveSetRevisionItem": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "event_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Date"
          },
          "ingested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ingested At"
          },
          "review_status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "file_count": {
            "type": "integer",
            "title": "File Count"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "source_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ref"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "title",
          "ingested_at",
          "review_status",
          "file_count"
        ],
        "title": "ArchiveSetRevisionItem"
      },
      "ArchiveSetsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ArchiveSetNode"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "total_sets": {
            "type": "integer",
            "title": "Total Sets"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated",
            "default": false
          },
          "max_documents_scanned": {
            "type": "integer",
            "title": "Max Documents Scanned"
          }
        },
        "type": "object",
        "required": [
          "page",
          "size",
          "total_sets",
          "generated_at",
          "max_documents_scanned"
        ],
        "title": "ArchiveSetsResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewStatus": {
        "type": "string",
        "enum": [
          "NONE",
          "NEEDS_REVIEW",
          "RESOLVED"
        ],
        "title": "ReviewStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}